+++ utils/postgre/config/config.js
@@ -1,7 +1,10 @@
 "use strict";
 require('dotenv').config();
+
+console.log("Config file: Test .env connection")
 console.log(process.env.PSQL_PASS)
 console.log(process.env.PSQL_USER)
+
 module.exports = {
   development: {
     username: process.env.PSQL_USER || "postgres",
