+++ utils/postgre/models/index.js
@@ -35,41 +35,18 @@ if (config.use_env_variable) {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
-
-//Import
+//Create Database Object
 db.receivers = require("./receivers.js")(sequelize, Sequelize.DataTypes);
 db.users = require("./users.js")(sequelize, Sequelize.DataTypes);
 
-//Associations
+//Create Associations
 db.users.hasMany(db.receivers, {
-  foreignKey: 'userId',
+  foreignKey: 'userID',
 });
 db.receivers.belongsTo(db.users, {
-  foreignKey: 'userId',
+  foreignKey: 'userID',
 });
-
-
-console.log("Hello")
-console.log(db.users.associations)
-console.log(db.receivers.associations)
-
-console.log("Hiiiiiiiiiii");
-console.log(Object.keys(db.users.associations));
-console.log(Object.keys(db.receivers.associations));
-
-
-// Object.keys(db).forEach(modelName => {
-//   console.log(db[modelName])
-//   if (db[modelName].associate) {
-//     db[modelName].associate(db);
-
-//     console.log(`Associations for ${modelName}:`);
-//     console.log(associations);
-//   } 
-//});
-
 
 //sequelize migration:generate --name receiverAddFogKey
 //sequelize seed:generate --name your-seed-file-name
-
-module.exports = db;
+module.exports =  db;
